+++ src/Domain/Item/SaveProcessor/ItemSaveProcessingDTO.ts
@@ -3,6 +3,7 @@ import { ItemHash } from '../ItemHash'
 
 export type ItemSaveProcessingDTO = {
   userUuid: string
+  apiVersion: string
   itemHash: ItemHash
   existingItem?: Item
 }
